+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2025-04-12
+
+### Added
+- Added command "Perfetto: Open Trace For File" to allow user to choose a trace file in workspace.
+
+### Changed
+- Renamed command "Perfetto: Open Trace" to "Perfetto: Open Trace For Active Editor".
+
 ## [0.1.0] - 2025-04-12
 
 ### Added